Added pagination and product page
Change Log:

1.Added Pagination,limiting products per page to 4 for now
2.Added product pages with hover over product to zoom.
3.Fixed some bugs related to layout in product page.
4.Removed time feature in admin side,fixed a bug in form that displayed incorrect field names in Addproduct form.
5.Updated bootstrap from v4 to v5 for better layout rendering.
6.Added carts folder and created carts.py file to store user carts
7.Created a new database.
8.Fixed render issue with admin registration.
9.Disabled email notif for now

Added more bugs to solve later

+++ shop/products/forms.py
@@ -3,7 +3,7 @@ from wtforms import Form,IntegerField,BooleanField,TextAreaField,validators,Stri
 
 class Addproducts(Form):
     name=StringField('',[validators.DataRequired()])
-    price=DecimalField('',[validators.DataRequired()])
+    price=IntegerField('',[validators.DataRequired()])
     discount=IntegerField('',)
     stock=IntegerField('',[validators.DataRequired()])
     origin=StringField('',[validators.DataRequired()])
